Fix context analysis & add shape_of instruction to vm

+++ python/tvm/relay/op/tensor.py
@@ -989,7 +989,7 @@ def device_copy(data, src_dev, dst_dev):
     return _make.device_copy(data, src_dev, dst_dev)
 
 
-def shape_of(data, dtype="int32"):
+def shape_of(data, dtype="int64"):
     """Get shape of a tensor.
 
     Parameters
